backend/requirements.txt
Flask
numpy
tensorflow
tensorflow_hub
Pillow
